Regroup bump unit tests by responsibility (#97)
Pure test reorganisation — no test logic, assertions, or Hypothesis
strategies changed. Replaces the earlier three-way split with five modules
grouped by the behaviour under test.

- test_bump_selectors.py: package-selector and skip helpers.
- test_bump_result_formatting.py: change-description counting, no-change /
  header / path formatting, the snapshot-backed result messages, and
  `_update_crate_manifest` with its co-located workspace-construction
  fixtures (UpdateCrateTestParams, _make_*_dependency, _parse_manifest_versions).
- test_bump_manifest_writing.py: `_update_manifest` and
  `_update_dependency_sections` rewrites.
- test_bump_toml_helpers.py: low-level `_select_table` / `_assign_version` /
  `_value_matches` helpers.
- test_bump_context_properties.py: unchanged issue #97 property tests.

Removes test_bump_command_internals.py, test_bump_manifest_editing.py, and
test_bump_dependency_sections.py. The syrupy snapshot file follows its tests
(renamed to test_bump_result_formatting.ambr). Each module carries only the
imports its tests use. All 29 test functions (30 items) collect and pass with
identical IDs; make check-fmt/lint/typecheck all green.
