--- tests/unit/test_bump_manifest_editing.py
@@ -1,203 +1,0 @@
-"""Unit tests for bump manifest editing: crate updates and TOML primitives.
-
-Covers :func:`lading.commands.bump._update_crate_manifest`,
-:func:`~lading.commands.bump._update_manifest`, and the low-level table and
-version helpers (``_select_table``, ``_assign_version``, ``_value_matches``).
-"""
-
-from __future__ import annotations
-
-from pathlib import Path
-
-import pytest
-from tomlkit import parse as parse_toml
-
-from lading.commands import bump
-from tests.helpers.workspace_builders import (
-    _load_version,
-    _make_config,
-)
-from tests.unit.test_bump_command_internals import (
-    UpdateCrateTestParams,
-    _make_workspace_with_alpha_dependency,
-)
-
-
-def _parse_manifest_versions(
-    manifest_path: Path,
-) -> tuple[str, str]:
-    """Extract package version and alpha dependency version from manifest.
-
-    Args:
-        manifest_path: Path to the Cargo.toml manifest.
-
-    Returns
-    -------
-        Tuple of (package_version, alpha_dependency_version).
-
-    """
-    document = parse_toml(manifest_path.read_text(encoding="utf-8"))
-    package_version = document["package"]["version"]
-    alpha_version = document["dependencies"]["alpha"].value
-    return package_version, alpha_version
-
-
-@pytest.mark.parametrize(
-    "params",
-    [
-        UpdateCrateTestParams(
-            test_id="excluded_crate_skips_version_bump",
-            dependency_spec=("alpha", "0.1.0"),
-            exclude_crates=("beta",),
-            expected_package_version="0.1.0",
-            expected_alpha_version="1.2.3",
-        ),
-        UpdateCrateTestParams(
-            test_id="updates_version_and_dependencies",
-            dependency_spec=("alpha", "^0.1.0"),
-            exclude_crates=(),
-            expected_package_version="1.2.3",
-            expected_alpha_version="^1.2.3",
-        ),
-    ],
-    ids=lambda p: p.test_id,
-)
-def test_update_crate_manifest(tmp_path: Path, params: UpdateCrateTestParams) -> None:
-    """Crate manifest updates handle exclusions and dependency rewrites."""
-    crate, workspace = _make_workspace_with_alpha_dependency(
-        tmp_path,
-        dependency=params.dependency_spec,
-    )
-    context = bump._initialize_bump_context(
-        tmp_path,
-        bump.BumpOptions(
-            configuration=_make_config(exclude=params.exclude_crates),
-            workspace=workspace,
-        ),
-    )
-
-    changed = bump._update_crate_manifest(
-        crate,
-        "1.2.3",
-        context,
-    )
-
-    assert changed is True
-    package_version, alpha_version = _parse_manifest_versions(crate.manifest_path)
-    assert package_version == params.expected_package_version
-    assert alpha_version == params.expected_alpha_version
-
-
-def test_update_manifest_writes_when_changed(tmp_path: Path) -> None:
-    """Applying a new version persists changes to disk."""
-    manifest_path = tmp_path / "Cargo.toml"
-    manifest_path.write_text('[package]\nname = "demo"\nversion = "0.1.0"\n')
-    changed = bump._update_manifest(
-        manifest_path, (("package",),), "1.0.0", bump.BumpOptions()
-    )
-    assert changed is True
-    assert _load_version(manifest_path, ("package",)) == "1.0.0"
-
-
-def test_update_manifest_preserves_inline_comment(tmp_path: Path) -> None:
-    """Inline comments survive manifest rewrites."""
-    manifest_path = tmp_path / "Cargo.toml"
-    manifest_path.write_text(
-        '[package]\nversion = "0.1.0"  # keep me\n', encoding="utf-8"
-    )
-    changed = bump._update_manifest(
-        manifest_path, (("package",),), "1.2.3", bump.BumpOptions()
-    )
-    assert changed is True
-    text = manifest_path.read_text(encoding="utf-8")
-    assert "# keep me" in text
-    document = parse_toml(text)
-    assert document["package"]["version"] == "1.2.3"
-
-
-def test_update_manifest_skips_when_unchanged(tmp_path: Path) -> None:
-    """No write occurs when the manifest already records the target version."""
-    manifest_path = tmp_path / "Cargo.toml"
-    original = '[package]\nname = "demo"\nversion = "0.1.0"\n'
-    manifest_path.write_text(original)
-    changed = bump._update_manifest(
-        manifest_path, (("package",),), "0.1.0", bump.BumpOptions()
-    )
-    assert changed is False
-    assert manifest_path.read_text() == original
-
-
-def test_select_table_returns_nested_table() -> None:
-    """Select nested tables using dotted selectors."""
-    document = parse_toml('[workspace]\n[workspace.package]\nversion = "0.1.0"\n')
-    table = bump._select_table(document, ("workspace", "package"))
-    assert table is document["workspace"]["package"]
-
-
-def test_select_table_returns_none_for_missing() -> None:
-    """Selectors that do not resolve to tables return ``None``."""
-    document = parse_toml("[workspace]\nmembers = []\n")
-    table = bump._select_table(document, ("workspace", "package"))
-    assert table is None
-
-
-def test_assign_version_handles_absent_table() -> None:
-    """``_assign_version`` tolerates missing tables."""
-    assert bump._assign_version(None, "1.0.0") is False
-
-
-def test_assign_version_updates_value() -> None:
-    """Assign a new version when the stored value differs."""
-    table = parse_toml('[package]\nname = "demo"\nversion = "0.1.0"\n')["package"]
-    assert bump._assign_version(table, "2.0.0") is True
-    assert table["version"] == "2.0.0"
-
-
-def test_assign_version_detects_existing_value() -> None:
-    """Return ``False`` when the version already matches."""
-    table = parse_toml('[package]\nversion = "0.1.0"\n')["package"]
-    assert bump._assign_version(table, "0.1.0") is False
-
-
-def test_value_matches_accepts_plain_strings() -> None:
-    """Strings compare directly when checking for version matches."""
-    assert bump._value_matches("1.0.0", "1.0.0") is True
-    assert bump._value_matches("1.0.0", "2.0.0") is False
-
-
-def test_value_matches_handles_toml_items() -> None:
-    """TOML items compare via their stored string value."""
-    document = parse_toml('version = "3.0.0"')
-    item = document["version"]
-    assert bump._value_matches(item, "3.0.0") is True
-    assert bump._value_matches(item, "4.0.0") is False
-
-
-def test_select_table_handles_out_of_order_package() -> None:
-    """Out-of-order tables (OutOfOrderTableProxy) are accepted."""
-    # When [package.metadata.docs.rs] appears after other tables,
-    # tomlkit returns an OutOfOrderTableProxy instead of a Table
-    document = parse_toml(
-        '[package]\nname = "x"\nversion = "0.1.0"\n'
-        "[dependencies]\n"
-        'foo = "1"\n'
-        "[package.metadata.docs.rs]\n"
-        "all-features = true\n"
-    )
-    table = bump._select_table(document, ("package",))
-    assert table is not None
-    assert table.get("version") == "0.1.0"
-
-
-def test_assign_version_works_with_out_of_order_table() -> None:
-    """Version assignment works with OutOfOrderTableProxy."""
-    document = parse_toml(
-        '[package]\nname = "x"\nversion = "0.1.0"\n'
-        "[dependencies]\n"
-        'foo = "1"\n'
-        "[package.metadata.docs.rs]\n"
-        "all-features = true\n"
-    )
-    table = bump._select_table(document, ("package",))
-    assert bump._assign_version(table, "2.0.0") is True
-    assert table.get("version") == "2.0.0"
